+++ src/map/Field.h
@@ -14,15 +14,15 @@ class Field {
     void draw() const;
     void reset_metrics();
     std::pair<int, int> get_pos() const;
-    double distance_to(Field& field) const;
+    int distance_to(Field& field) const;
     bool is_empty() const;
     bool is_walkable() const;
     bool operator==(const Field& rhs) const;
 
-    double distance{BASE_METRIC_VALUE};
-    double ca_dist{BASE_METRIC_VALUE};
-    double ch_dist{BASE_METRIC_VALUE};
-    double fo_dist{BASE_METRIC_VALUE};
+    int distance{BASE_METRIC_VALUE};
+    int ca_dist{BASE_METRIC_VALUE};
+    int ch_dist{BASE_METRIC_VALUE};
+    int fo_dist{BASE_METRIC_VALUE};
 
    protected:
     int m_x;
